Extract hook for default shipping method
